chore: Move builder code from pallets to src folder 🗂️
