+++ src/builder/builders/Builder.test.ts
@@ -2,13 +2,13 @@
 
 import { type ApiPromise } from '@polkadot/api'
 import { vi, describe, expect, it, beforeEach } from 'vitest'
-import { type Bool, type TNode } from '../../../types'
-import { createApiInstance } from '../../../utils'
-import * as hrmp from '../../hrmp'
-import * as parasSudoWrapper from '../../parasSudoWrapper'
-import * as xcmPallet from '../../xcmPallet'
-import * as xyk from '../../xyk'
-import { getRelayChainSymbol } from '../../assets'
+import { type Bool, type TNode } from '../../types'
+import { createApiInstance } from '../../utils'
+import * as hrmp from '../../pallets/hrmp'
+import * as parasSudoWrapper from '../../pallets/parasSudoWrapper'
+import * as xcmPallet from '../../pallets/xcmPallet'
+import * as xyk from '../../pallets/xyk'
+import { getRelayChainSymbol } from '../../pallets/assets'
 import { Builder } from './Builder'
 
 const WS_URL = 'wss://para.f3joule.space'
