+++ src/index.ts
@@ -6,8 +6,8 @@ export * as openChannels from './pallets/parasSudoWrapper'
 export * as closeChannels from './pallets/hrmp'
 export * as assets from './pallets/assets'
 export * from './pallets/assets'
-export * from './pallets/builder'
 export * from './pallets/pallets'
+export * from './builder'
 export * from './types'
 export { NODE_NAMES, NODES_WITH_RELAY_CHAINS, SUPPORTED_PALLETS } from './maps/consts'
 export {
